Move Resource package to SDK
