+++ sdk/src/test/java/io/opentelemetry/sdk/resource/ResourceTest.java
@@ -14,11 +14,12 @@
  * limitations under the License.
  */
 
-package io.opentelemetry.resources;
+package io.opentelemetry.sdk.resource;
 
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.testing.EqualsTester;
+import io.opentelemetry.sdk.resources.Resource;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
